# SwaggerOffLine.yaml
openapi: 3.1.0
info:
  title: Auto Escola 3ESPG
  description: API REST para gerenciamento de instruções da Auto Escola 3ESPG.
  contact:
    name: Auto Escola 3ESPG
    email: autoescola3espg@email.com
  license:
    name: Apache 2.0
    url: http://springdoc.org
servers:
  - url: http://localhost:8085
    description: Generated server url
paths:
  /instrutores:
    get:
      tags:
        - instrutor-controller
      operationId: listarInstrutores
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
        - name: sort
          in: query
          description: "Sorting criteria in the format: property,(asc|desc). Default sort
            order is ascending. Multiple sort criteria are supported."
          required: false
          schema:
            type: array
            default:
              - nome,ASC
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/PageDadosListagemInstrutor"
      security:
        - bearer-key: []
    put:
      tags:
        - instrutor-controller
      operationId: atualizarInstrutor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAtualizacaoInstrutor"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
      security:
        - bearer-key: []
    post:
      tags:
        - instrutor-controller
      operationId: cadastrarInstrutor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroInstrutor"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
      security:
        - bearer-key: []
  /login:
    post:
      tags:
        - login-controller
      operationId: efetuarLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosLogin"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: string
  /instrucoes:
    post:
      tags:
        - instrucao-controller
      operationId: agendarInstrucao
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAgendamentoInstrucao"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAgendamento"
  /instrutores/{id}:
    get:
      tags:
        - instrutor-controller
      operationId: detalharInstrutor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
      security:
        - bearer-key: []
    delete:
      tags:
        - instrutor-controller
      operationId: excluirInstrutor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
      security:
        - bearer-key: []
  /health-check:
    get:
      tags:
        - health-check-controller
      operationId: healthCheck
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: string
components:
  schemas:
    DadosAtualizacaoInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
      required:
        - id
    DadosEndereco:
      type: object
      properties:
        logradouro:
          type: string
          minLength: 1
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
          minLength: 1
        cidade:
          type: string
          minLength: 1
        uf:
          type: string
          minLength: 1
          pattern: "[A-Z]{2}"
        cep:
          type: string
          minLength: 1
          pattern: "[0-9]{5}-[0-9]{3}"
      required:
        - bairro
        - cep
        - cidade
        - logradouro
        - uf
    DadosDetalhamentoInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        cnh:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
        ativo:
          type: boolean
    DadosLogin:
      type: object
      properties:
        login:
          type: string
          minLength: 1
        senha:
          type: string
          minLength: 1
      required:
        - login
        - senha
    DadosCadastroInstrutor:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        telefone:
          type: string
          minLength: 1
        cnh:
          type: string
          minLength: 1
          pattern: \d{9,11}
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
      required:
        - cnh
        - email
        - especialidade
        - nome
        - telefone
    DadosAgendamentoInstrucao:
      type: object
      properties:
        id_aluno:
          type: integer
          format: int64
        id_instrutor:
          type: integer
          format: int64
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        data_hora:
          type: string
          format: date-time
      required:
        - data_hora
        - id_aluno
    DadosDetalhamentoAgendamento:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nomeAluno:
          type: string
        nomeInstrutor:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        dataHora:
          type: string
          format: date-time
    DadosListagemInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
    PageDadosListagemInstrutor:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosListagemInstrutor"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT