# src/main/java/br/com/fiap3espg/autoescola3espg/domain/usuario/DadosLogin.java
package br.com.fiap3espg.autoescola3espg.domain.usuario;

import jakarta.validation.constraints.NotBlank;

public record DadosLogin(
        @NotBlank
        String login,

        @NotBlank
        String senha) {
}